+++ src/backend/win_cid/file_dialog/dialog_ffi.rs
@@ -12,7 +12,7 @@ use windows_sys::{
         System::Com::{CoCreateInstance, CLSCTX_INPROC_SERVER},
         UI::Shell::{
             FileOpenDialog, FileSaveDialog, SHCreateItemFromParsingName, FOS_ALLOWMULTISELECT,
-            FOS_PICKFOLDERS,
+            FOS_FORCESHOWHIDDEN, FOS_PICKFOLDERS,
         },
     },
 };
@@ -79,6 +79,20 @@ impl DialogInner {
     unsafe fn set_options(&self, opts: FILEOPENDIALOGOPTIONS) -> Result<()> {
         let (d, v) = self.fd();
         wrap_err((v.SetOptions)(d, opts))
+    }
+
+    #[inline]
+    unsafe fn get_options(&self) -> Result<FILEOPENDIALOGOPTIONS> {
+        let (d, v) = self.fd();
+        let mut opts = 0;
+        wrap_err((v.GetOptions)(d, &mut opts))?;
+        Ok(opts)
+    }
+
+    #[inline]
+    unsafe fn add_options(&self, opts: FILEOPENDIALOGOPTIONS) -> Result<()> {
+        let current = self.get_options()?;
+        self.set_options(current | opts)
     }
 
     #[inline]
@@ -273,6 +287,15 @@ impl IDialog {
         Ok(())
     }
 
+    fn set_show_hidden_files(&self, show: Option<bool>) -> Result<()> {
+        if let Some(true) = show {
+            unsafe {
+                self.0.add_options(FOS_FORCESHOWHIDDEN)?;
+            }
+        }
+        Ok(())
+    }
+
     pub fn get_results(&self) -> Result<Vec<PathBuf>> {
         unsafe { self.0.get_results() }
     }
@@ -294,6 +317,7 @@ impl IDialog {
         dialog.set_path(&opt.starting_directory)?;
         dialog.set_file_name(&opt.file_name)?;
         dialog.set_title(&opt.title)?;
+        dialog.set_show_hidden_files(opt.show_hidden_files)?;
 
         Ok(dialog)
     }
@@ -305,6 +329,7 @@ impl IDialog {
         dialog.set_path(&opt.starting_directory)?;
         dialog.set_file_name(&opt.file_name)?;
         dialog.set_title(&opt.title)?;
+        dialog.set_show_hidden_files(opt.show_hidden_files)?;
 
         Ok(dialog)
     }
@@ -315,8 +340,13 @@ impl IDialog {
         dialog.set_path(&opt.starting_directory)?;
         dialog.set_title(&opt.title)?;
 
+        let mut opts = FOS_PICKFOLDERS;
+        if let Some(true) = opt.show_hidden_files {
+            opts |= FOS_FORCESHOWHIDDEN;
+        }
+
         unsafe {
-            dialog.0.set_options(FOS_PICKFOLDERS)?;
+            dialog.0.set_options(opts)?;
         }
 
         Ok(dialog)
@@ -327,7 +357,11 @@ impl IDialog {
 
         dialog.set_path(&opt.starting_directory)?;
         dialog.set_title(&opt.title)?;
-        let opts = FOS_PICKFOLDERS | FOS_ALLOWMULTISELECT;
+
+        let mut opts = FOS_PICKFOLDERS | FOS_ALLOWMULTISELECT;
+        if let Some(true) = opt.show_hidden_files {
+            opts |= FOS_FORCESHOWHIDDEN;
+        }
 
         unsafe {
             dialog.0.set_options(opts)?;
@@ -344,8 +378,13 @@ impl IDialog {
         dialog.set_file_name(&opt.file_name)?;
         dialog.set_title(&opt.title)?;
 
+        let mut opts = FOS_ALLOWMULTISELECT;
+        if let Some(true) = opt.show_hidden_files {
+            opts |= FOS_FORCESHOWHIDDEN;
+        }
+
         unsafe {
-            dialog.0.set_options(FOS_ALLOWMULTISELECT)?;
+            dialog.0.set_options(opts)?;
         }
 
         Ok(dialog)